+++ SRP/Dao/UserManager.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace SRP\Dao;
+
+use SRP\User;
+
+class UserManager
+{
+    public static function insert(User $user)
+    {
+
+    }
+
+    public static function update(User $user)
+    {
+
+    }
+}
